+++ crates/core/action-pipeline/src/actions/run_target.rs
@@ -3,11 +3,8 @@ use moon_action::{Action, ActionStatus};
 use moon_action_context::ActionContext;
 use moon_emitter::Emitter;
 use moon_logger::{color, debug, warn};
-use moon_node_platform::actions as node_actions;
-use moon_platform::PlatformType;
 use moon_project::Project;
-use moon_runner::{HydrateFrom, Runner};
-use moon_system_platform::actions as system_actions;
+use moon_runner::Runner;
 use moon_task::Target;
 use moon_terminal::Checkpoint;
 use moon_workspace::Workspace;
@@ -65,58 +62,9 @@ pub async fn run_target(
     // Abort early if this build has already been cached/hashed
     if should_cache {
         let mut context = context.write().await;
-        let common_hasher = runner.create_common_hasher(&context).await?;
 
-        let is_cached = match task.platform {
-            PlatformType::Node => {
-                runner
-                    .is_cached(
-                        &mut context,
-                        common_hasher,
-                        node_actions::create_target_hasher(&workspace, project).await?,
-                    )
-                    .await?
-            }
-            PlatformType::System => {
-                runner
-                    .is_cached(
-                        &mut context,
-                        common_hasher,
-                        system_actions::create_target_hasher(&workspace, project)?,
-                    )
-                    .await?
-            }
-            PlatformType::Unknown => None,
-        };
-
-        if let Some(cache_location) = is_cached {
-            // Only hydrate when the hash is different from the previous build,
-            // as we can assume the outputs from the previous build still exist?
-            if matches!(cache_location, HydrateFrom::LocalCache)
-                || matches!(cache_location, HydrateFrom::RemoteCache)
-            {
-                runner.hydrate_outputs().await?;
-            }
-
-            let mut comments = vec![match cache_location {
-                HydrateFrom::LocalCache => "cached",
-                HydrateFrom::RemoteCache => "cached from remote",
-                HydrateFrom::PreviousOutput => "cached from previous run",
-            }];
-
-            if runner.should_print_short_hash() {
-                comments.push(runner.get_short_hash());
-            }
-
-            runner.print_checkpoint(Checkpoint::RunPassed, &comments)?;
-            runner.print_cache_item()?;
-            runner.flush_output()?;
-
-            return Ok(if matches!(cache_location, HydrateFrom::RemoteCache) {
-                ActionStatus::CachedFromRemote
-            } else {
-                ActionStatus::Cached
-            });
+        if let Some(cache_location) = runner.is_cached(&mut context).await? {
+            return Ok(runner.hydrate(cache_location).await?);
         }
     }