+++ crates/system/platform/src/lib.rs
@@ -3,5 +3,4 @@ mod hasher;
 mod platform;
 mod tool;
 
-pub use hasher::SystemTargetHasher;
 pub use platform::SystemPlatform;